refactor: migrate fieldSettings:TableColumn
